requirements.txt
aiofiles==0.5.0
aiosqlite==0.16.1
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
asyncpg==0.24.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.68.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
httptools==0.1.2
idna==3.2
iniconfig==1.1.1
iso8601==0.1.16
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
orjson==3.6.1
packaging==21.0
pluggy==0.13.1
promise==2.3
py==1.10.0
pydantic==1.8.2
pyparsing==2.4.7
pypika-tortoise==0.1.1
pytest==6.2.4
pytest-asyncio==0.15.1
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
Rx==1.6.1
six==1.16.0
starlette==0.14.2
toml==0.10.2
tortoise-orm==0.17.6
typing-extensions==3.10.0.0
ujson==4.1.0
urllib3==1.26.6
uvicorn==0.13.4
uvloop==0.16.0
watchgod==0.7
websockets==8.1
